refactor: import useGroupDAO from context
