+++ src/ProposalList.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useMemo } from "react";
-import { useGroupDAO } from "./App";
+import { useGroupDAO } from "./GroupDAOContext";
 import { Shield, ThumbsUp, ThumbsDown, CircleSlash, Loader2 } from "lucide-react";
 import { toast } from "react-toastify";
 import { format } from "date-fns";
